@@ -1,12 +1,12 @@
 name: japhy_todo_app
 description: A Flutter To-Do app called Japhy.
 
-publish_to: 'none' # Remove this line if you plan to publish to pub.dev
+publish_to: 'none'
 
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=2.19.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -14,16 +14,23 @@ dependencies:
   cupertino_icons: ^1.0.2
   provider: ^6.0.5
   table_calendar: ^3.0.9
-  firebase_core: ^2.10.0
-  firebase_auth: ^4.4.0
-  flutter_local_notifications: ^13.0.0
-  intl: ^0.18.1
-  timezone: ^0.9.0
+  firebase_core: ^3.14.0
+  firebase_auth: ^5.6.0
+  cloud_firestore: ^5.6.9
+  flutter_local_notifications: ^19.3.0
+  intl: ^0.20.2
+  timezone: ^0.10.1
+  uuid: ^4.3.1
+
+flutter_local_notifications:
+  default_os_configs:
+    linux:
+      default_package: null
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^6.0.0
 
 flutter:
   uses-material-design: true
